file_parser/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
openpyxl>=3.1.0
Pillow>=9.0.0
python-decouple>=3.8 